[BBC-9070] MessagingSummaryItem.MDX
