+++ src/messagingSummaryItem/MessagingSummaryItem.style.tsx
@@ -1,3 +1,14 @@
 import styled from 'styled-components'
 
-export const StyledMessagingSummaryItem = styled.div``
+import { Item } from '../_internals/item'
+
+export const StyledMessagingSummaryItem = styled(Item)`
+  & .kirk-messaging-summary-item-sub-label {
+    display: flex;
+    word-break: break-word;
+
+    /* Truncate sublabel to max 2 lines. */
+    max-height: 2.5em;
+    overflow: hidden;
+  }
+`
